+++ scripts/start-nomad.sh
@@ -20,17 +20,49 @@ if [ ! -e /tmp/consul-test.pid ]; then
 fi
 
 if [ ! -e /tmp/nomad-test.pid ]; then
-    nomad agent -dev -acl-enabled -vault-address=$VAULT_ADDR -vault-token $VAULT_TEST_TOKEN -vault-enabled -vault-allow-unauthenticated=false > /tmp/nomad.log 2>&1 &
+    cat <<EOF > /tmp/nomad-config.hcl
+log_file = "/tmp/nomad.log"
+
+plugin "docker" {
+  config {
+    allow_privileged = true
+  }
+}
+EOF
+
+    sudo -Eb nomad agent -dev -acl-enabled \
+      -data-dir=/tmp/nomad/data \
+      -config=/tmp/nomad-config.hcl \
+      -vault-address=$VAULT_ADDR \
+      -vault-token=$VAULT_TEST_TOKEN \
+      -vault-enabled \
+      -vault-allow-unauthenticated=false
     NOMAD_PID=$!
     echo $NOMAD_PID > /tmp/nomad-test.pid
 
     # Give some time for the process to initialize
     sleep 10
 
-    http --ignore-stdin POST http://localhost:4646/v1/acl/bootstrap | jq -r '.SecretID' > /tmp/nomad-test.token
-    NOMAD_TOKEN=$(cat /tmp/nomad-test.token)
+    nomad operator api -X POST /v1/acl/bootstrap | jq -r '.SecretID' > /tmp/nomad-test.token
+    export NOMAD_TOKEN=$(cat /tmp/nomad-test.token)
+    if [ -z "$NOMAD_TOKEN" ]; then
+      echo "Failed to bootstrap Nomad ACL"
+      exit 1
+    fi
     echo "NOMAD_TOKEN=$(echo $NOMAD_TOKEN)" >> $GITHUB_ENV
-elif [ -e /tmp/nomad-test.token ]; then 
+
+    # Run hostpath CSI plugin and wait for it to be healthy.
+    nomad job run https://raw.githubusercontent.com/hashicorp/nomad/v1.3.1/demo/csi/hostpath/plugin.nomad
+    echo "Waiting for hostpath CSI plugin to become healthy"
+    retries=30
+    while [ $retries -ge 0 ]; do
+        nomad plugin status hostpath \
+            | grep "Nodes Healthy        = 1" && break
+        sleep 2
+        retries=$(( $retries - 1 ))
+    done
+    nomad plugin status hostpath
+elif [ -e /tmp/nomad-test.token ]; then
   NOMAD_TOKEN=$(cat /tmp/nomad-test.token)
   echo "NOMAD_TOKEN=$(echo $NOMAD_TOKEN)" >> $GITHUB_ENV
 fi